refactor: organise node classes into packages
